+++ genium/src/test/resources/smoke/external_types/output/cpp/src/smoke/FooClass.cpp
@@ -10,16 +10,16 @@ namespace smoke {
 static_assert(
     std::is_same<
         ::std::string,
-        decltype(((::foo::BarClass*)nullptr)->doStuff(
-            *((int32_t*)nullptr)))
+        std::remove_cv<std::remove_reference<decltype(((::foo::BarClass*)nullptr)->doStuff(
+            *((int32_t*)nullptr)))>::type>::type
     >::value,
     "Expected '::std::string' return type for '::foo::BarClass::doStuff'."
 );
 static_assert(
     std::is_same<
         ::foo::BarEnum,
-        decltype(::foo::BarClass::doStaticStuff(
-            *((::foo::BarEnum*)nullptr)))
+        std::remove_cv<std::remove_reference<decltype(::foo::BarClass::doStaticStuff(
+            *((::foo::BarEnum*)nullptr)))>::type>::type
     >::value,
     "Expected '::foo::BarEnum' return type for '::foo::BarClass::doStaticStuff'."
 );